+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleopAll.java
@@ -1,8 +1,11 @@
 package org.firstinspires.ftc.teamcode;
 
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+import org.firstinspires.ftc.teamcode.conveyor.ConveyorCommon;
+import org.firstinspires.ftc.teamcode.drivetrain.DrivetrainCommon;
+import org.firstinspires.ftc.teamcode.spinner.SpinnerCommon;
 
 @TeleOp(name="Teleop All", group="Pushbot")
 //@Disabled
