+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autons/AutoHardware.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autons;
 
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
