+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drivetrain/DrivetrainTeleop.java
@@ -1,9 +1,11 @@
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.drivetrain;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+
+import org.firstinspires.ftc.teamcode.drivetrain.DrivetrainCommon;
 
 @TeleOp(name="Drivetrain Teleop")
 @Disabled
