+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/extender/ExtenderCommon.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.extender;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.Servo;
 
 public class ExtenderCommon {
