+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/extender/ExtenderHardware.java
@@ -1,21 +1,24 @@
 /*
-
 This is the Hardware class, it is used to declare all the components and create objects
 of them so they can be used in the code to run them
-
  */
 
 // This just tells the code where this file is located in the file structure
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.extender;
 
 // This imports all the different dependencies and libraries and other classes we use in this file
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-public class SpinnerHardware {
-    public DcMotor spinnerMotor = null;
+public class ExtenderHardware {
+
+    public Servo s1;
+    public Servo s2;
+    public Servo s3;
+    public Servo s4;
+    public Servo s5;
 
     HardwareMap hwMap =  null;
     private ElapsedTime period = new ElapsedTime();
@@ -23,13 +26,10 @@ public class SpinnerHardware {
     public void init(HardwareMap ahwMap) {
         hwMap = ahwMap;
 
-        spinnerMotor  = hwMap.get(DcMotor.class, "spinner_motor");
-
-        spinnerMotor.setDirection(DcMotor.Direction.FORWARD);
-
-        spinnerMotor.setPower(0);
-
-        spinnerMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        s1 = hwMap.get(Servo.class, "s_1");
+        s2 = hwMap.get(Servo.class, "s_2");
+        s3 = hwMap.get(Servo.class, "s_3");
+        s4 = hwMap.get(Servo.class, "s_4");
+        s5 = hwMap.get(Servo.class, "s_5");
     }
- }
-
+}
