+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/intake/IntakeTeleop.java
@@ -1,8 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.intake;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+import org.firstinspires.ftc.teamcode.intake.IntakeCommon;
 
 @TeleOp(name="Intake Teleop", group="Pushbot")
 @Disabled
