+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/olddrivetrain/OldDrivetrainCommon.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.olddrivetrain;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.HardwareMap;
 
 public class OldDrivetrainCommon {
