+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/spinner/SpinnerCommon.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.spinner;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
